--- src/components/Header/MainHeader/MainHeader.module.scss
@@ -1,100 +1,0 @@
-.mainHeader {
-  width: 1280px;
-  padding: 23px 0px;
-  gap: 24px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  &_back {
-    background-color: #404966;
-  }
-
-  .btn {
-    width: 163px;
-    height: 48px;
-    display: flex;
-    padding: 8px;
-    justify-content: center;
-    align-items: center;
-    border-radius: 8px;
-    background: #cb5980;
-    font-size: 16px;
-    font-weight: 600;
-    color: $primary-text-color;
-    border: none;
-    gap: 8px;
-
-    &:hover {
-      background: #ef7da4;
-    }
-
-    &:focus {
-      background: #cb5980;
-    }
-  }
-
-  .inputControl {
-    max-width: 410px;
-    width: 100%;
-  }
-
-  .searchInput {
-    padding: 0 12px 0 40px;
-    color: $input-text-color;
-  }
-
-  .headerBlock {
-    @include flex-container;
-  }
-
-  .headerBlock:first-child {
-    gap: 100px;
-  }
-
-  .headerBlock:last-child {
-    flex: 1;
-  }
-  .svg {
-    width: 24px;
-    height: 24px;
-    fill: currentColor;
-  }
-  .svgSearch {
-    width: 24px;
-    height: 24px;
-    fill: transparent;
-    stroke: rgba(68, 68, 68, 1);
-  }
-  .svgClose {
-    width: 24px;
-    height: 24px;
-    fill: transparent;
-    stroke: currentColor;
-    transition: $transition;
-  }
-  .buttonSVG {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    cursor: pointer;
-    color: $input-text-color;
-    background: transparent;
-    border: none;
-    &:hover,
-    &:focus {
-      transform: scale(1.1);
-    }
-  }
-  .svgClose:hover,
-  .svgClose:focus {
-    transform: scale(1.1);
-  }
-  .buttonSearch {
-    background: transparent;
-    border: none;
-    outline: none;
-    cursor: pointer;
-  }
-}
