--- src/components/Header/SubHeader/SubHeader.module.scss
@@ -1,23 +1,0 @@
-.subHeader {
-  background-color: $nav-color;
-
-  .wrapper {
-    padding: 12px 10px;
-    @include flex-container;
-  }
-
-  .info {
-    @include flex-container;
-    gap: 26px;
-  }
-
-  .links {
-    @include flex-container;
-    gap: 16px;
-  }
-
-  .location {
-    @include flex-container;
-    gap: 4px;
-  }
-}
